move vlm_convert_model
